Incorrect Commit message. Based on the provided diff, a more appropriate commit message would be:
Add USERID argument in Dockerfile

Introduced USERID as an argument in Dockerfile to ensure flexibility and modularity. Instead of hardcoding a specific user id, this change allows the user id to be passed as an argument while building the Docker image. This will particularly be helpful in environments where user ids are managed centrally or follow a certain pattern.